requirements.txt
fasttext-wheel
transformers
torch
indic-transliteration